+++ src/core/config.h
@@ -7,11 +7,11 @@ namespace swarmnet_sim {
 class Sim_config {
    public:
     int get_num_robots() const;
-    int get_arena_max_x() const;
-    int get_arena_max_y() const;
+    double get_arena_max_x() const;
+    double get_arena_max_y() const;
     int get_rand_seed() const;
     // int get_ticks_per_second() const;
-    int get_duration() const;
+    double get_duration() const;
     int get_log_buf_size() const;
     void *get_robot_placement_dl_handle() const;
     void *get_robot_program_dl_handle() const;
@@ -20,11 +20,11 @@ class Sim_config {
     std::string get_motion_log_name() const;
 
     void set_num_robots(int val);
-    void set_arena_max_x(int val);
-    void set_arena_max_y(int val);
+    void set_arena_max_x(double val);
+    void set_arena_max_y(double val);
     void set_rand_seed(int val);
     // void set_ticks_per_second(int val);
-    void set_duration(int val);
+    void set_duration(double val);
     void set_log_buf_size(int val);
     void set_robot_placement_dl_handle(void *val);
     void set_robot_program_dl_handle(void *val);
@@ -34,11 +34,11 @@ class Sim_config {
 
    private:
     int num_robots;
-    int arena_max_x;
-    int arena_max_y;
+    double arena_max_x;
+    double arena_max_y;
     int rand_seed;
     // int ticks_per_second;
-    int duration;
+    double duration;
     int log_buf_size;
     void *robot_placement_dl_handle;
     void *robot_program_dl_handle;
