+++ src/plugin/engine/default_engine/default_engine.cpp
@@ -130,8 +130,8 @@ bool Default_engine::check_robot_collision(position2d_t pos_1,
     }
 }
 
-bool Default_engine::check_out_of_bound(position2d_t pos, int radius, int x_max,
-                                        int y_max) {
+bool Default_engine::check_out_of_bound(position2d_t pos, double radius,
+                                        double x_max, double y_max) {
     double x = pos.x;
     double y = pos.y;
