+++ src/plugin/engine/improved_analytical_engine/improved_analytical_engine.h
@@ -14,10 +14,10 @@ Physics_engine* engine_builder(void* arena);
 class Analytical_engine : public Physics_engine {
    public:
     double check_collision(double future_time);
-    double time_to_out_of_bound(position2d_t pos, double v, int radius,
-                                int x_max, int y_max);
+    double time_to_out_of_bound(position2d_t pos, double v, double radius,
+                                double x_max, double y_max);
     double time_to_collide(position2d_t pos_1, double v_1, position2d_t pos_2,
-                           double v_2, int radius);
+                           double v_2, double radius);
     void init();
     Analytical_engine(void* arena);
